rename: 폴더컨벤션 변경에따른 폴더명 변경
